refactor(invitation): colocate editor components
