+++ src/app/(main)/(my-order)/my-orders/[orderId]/invitation/_components/BasicInfoSection.tsx
@@ -2,7 +2,8 @@
 
 import { Card, CardContent, CardHeader, CardTitle } from "@/ui/components/atoms";
 import { format } from "date-fns";
-import { ComboboxField, DateField } from "@/ui/components/molecules";
+import { ComboboxField } from "./ComboboxField";
+import { DateField } from "./DateField";
 import { AddressField, SwitchField, TextField } from "@/ui/components/organisms";
 
 import type { InvitationContent } from "@/core/domain";
