+++ src/app/(main)/(my-order)/my-orders/[orderId]/invitation/_components/index.tsx
@@ -1,2 +1,1 @@
-export * from "./InvitationForm";
 export * from "./InvitationStatusControls";
