+++ src/ui/components/molecules/index.ts
@@ -3,9 +3,7 @@ export * from "./AutoCompleteList";
 export * from "./BankField";
 export * from "./BaseSelect";
 export * from "./CloudImage";
-export * from "./ComboboxField";
 export * from "./CursorPagination";
-export * from "./DateField";
 export * from "./LegalDocumentTemplate";
 export * from "./Pricing";
 export * from "./ProductCard";
