+++ src/ui/components/organisms/index.ts
@@ -1,16 +1,11 @@
 export * from "./AddressField";
-export * from "./BasicInfoSection";
 export * from "./BottomActionBar";
 export * from "./ClipboardButton";
 export * from "./ComingSoonTemplate";
-export * from "./InvitationFormView";
-export * from "./CoupleInfoSection";
 export * from "./ErrorFallback";
 export * from "./FormField";
 export * from "./ImageField";
 export * from "./ImagePreviewItem";
-export * from "./ImagesSection";
-export * from "./ParentsInfoSection";
 export * from "./PersonValueCard";
 export * from "./ProductCatalog";
 export * from "./ProductFeatures";
